project_output/requirements.txt
torch
torchvision
pandas
pillow
accelerate
